Field name 'description' from eventos is now 'event_data'.
Reason of change? We should store the iCal format specified in
the RFC of each event in this field. In case of presence of a
comment or description, the event_data will provide this info
only if we need it, as well as the rest of the event info.
The use of the fields on eventos should be
only for better performance on searchs involving recurrent events.